+++ store/index.js
@@ -2,7 +2,7 @@ import {applyMiddleware, createStore} from 'redux';
 import reducers from './reducers';
 import ReduxThunk from 'redux-thunk';
 
-export const makeStore = (initialState, options) => createStore(reducers,
+export const initializeStore = (initialState, options) => createStore(reducers,
   initialState,
   applyMiddleware(ReduxThunk)
 );
